Implemented WAITWMR flag; various notes updates.
 Changes to be committed:
	modified:   linux/graphics.c

1. Implemented WAITWMR flag, wait for window manager response. This flag
bypasses waiting until the window manager signals complete on certain
operations. See notes for more.

	modified:   linux/graphics.not
	modified:   portable/gnome_widgets.not

1. Various notes updates.

	modified:   tests/management_test.c
	modified:   tests/widget_test.c

Notes:

The need to wait for an Xwindow window manager response is an old issue with
this implementation. The Xwindow protocol should be logically ordered, for
instance if the size of a window is changed, further operations should apply to
the new window size.

It was determined fairly quickly that this is not so with Xwindows. The simple
proof was that a set of the window size, followed by reading the size, didn't
give the new size! For some operations, like size, I was able to simply store
new size locally, and read that back.

The theory was that the wait for window manager confimation was no longer
necessary, so the WAITWMR flag was implemented. However, although it passes most
tests, it did fail in widget_test/terminal number select box test, so it needs
further work.

It would also be helpful to run benchmarks and see if it actually makes a
significant difference in runtime.