Check hover active on focus.
 Changes to be committed:
	modified:   portable/managerc.c

1. In intevent(), when focus is entered on mouse click, also check for hover.
This is because we condition hover on focus, and the mouse will already have
moved.

Notes:

There is a condition where hover can escape the window. If the subwindow is
clipped against the root window edge, the mouse will not cross the frame and
remove hover. There doesen't seem to be anything that can be done for this. Even
if we cheat and use root interface hover to signal that, it won't work on a
terminal surface.