Moved foreground, background and attribute sets to globals.
 Changes to be committed:
	modified:   linux/terminal.c

1. Moved forec, backc, and attr out of screen data into globals.

	modified:   linux/terminal.not

2. Updated notes.

Notes:

This is a partial for the fix that all colors, attributes and other data is kept
globally and not in each screen. This matches other modules in the system, and
it makes sense as a simpler system. Although each screen buffer keeps track of
what was written there, what mode sets are done are true for all surfaces.

We have several other screen data items to give this treatment.