Fix winch/resize; add managerc notes.
Theme:

system_event.c was extensively refactored because the signal capture system
didn't work, to wit, pselect() does not bail for a winch signal (I think the
documentation was talking about SIGINT). The solution was to use signalfd() to
convert the signal to a file, and multiplex that using pselect().

 Changes to be committed:
	modified:   include/terminal.h

1. Add resize to terminal.h with "extra" parameters.

	modified:   linux/system_event.c

1. Add signalfd header.

2. Add PRTSEVT flag.

3. Remove sig_handler (unused now).

4. Refactor system_event_addsesig() to use signalfd().

5. Refactor system_event_getsevt() to treat signals as another fileofiled
event like timers.

	modified:   linux/terminal.c

1. Refactor winch/resize message.

	modified:   portable/managerc.c

1. Don't print graphical parameters to resize (cause character mode and all).

	modified:   portable/managerc.not

1. Add more found bugs (yea!).