Refactor restore() to handle resizes.
 Changes to be committed:
	modified:   linux/terminal.c

1. In restore(), set buffer size as clipped by actual screen image.

2. Clear right and bottom of screen beyond buffer.

Notes:

Some of this is to handle the specifics of xterm. For example, a screen
clear to xterm does the expected thing, which is clear the background to the
background color. However, a resize of the xterm window colors the "discovered"
area to xterms preset background, so we have to manually recover that.