Fix pselect() change issue.
With the move to events read on a thread, there is an issue that changes to the
pselect() mask like registering a timer, signal or fid don't work because the
thread is already in the pselect() call. So we make a user defined signal that
causes pselect() to exit and reenter, thus resetting it's mask.

 Changes to be committed:
	modified:   linux/system_event.c

1. Add reset system id resetsev, presently unused.

2. In system_event_addseinp(), send reset event.

3. In system_event_addsesig(), send reset event.

4. In system_event_addsetim(), send reset event.

5. In init(), set SIGUSR1 to ignore, and then add it to watched signals.
Apparently ignoring a signal does not mean the same thing as not signaling it
on an I/O channel. Linux is werid.