Added focus and nofocus messages to terminal.
 Changes to be committed:
	modified:   include/graphics.h
	modified:   include/terminal.h

1. Move focus, nofocus, hover and nohover up into the terminal set.

	modified:   linux/terminal.c

1. Add focus, nofocus, hover and nohover key equivalents to keytab[]. The hover
functions don't exist in xterm.

2. Add focus, nofocus, hover and nohover to prtevtt().

3. Enable focusin and focusout in init().

4. Disable focusin and focusout in deinit().

Notes:

focus, nofocus, hover and nohover are really windows management messages, but,
as in the (infamous) buffer resize contraversy, the answer for fixed terminal
functions that happen in a windowed environment is to use the features of
managed windows and bring them back into terminal on a "transparent" basis.

The focus and hover messages are marginally useful in terminal. Really the hover
message would have been the most useful, since hover edge escapes are a thing
in managementc. Focus is also useful, since it can be used to modulate the
cursor in managementc.

The moral of that story is for all packages to treat the focus and hover
messages as completely optional, and have a workaround planned for their
absense.