Added wrtstrn() to apis; Added BACKTRAP to stdio; Move dbg_prints to include; allow enter to exit autohold; add notes.
 Changes to be committed:
	modified:   include/graphics.h
	modified:   linux/graphics.c

1. Added wrtstrn() call from terminal.

2. Moved sizbug() up to terminal block in API list (no net functional
difference).

	modified:   libc/stdio.c
	modified:   libc/stdio.h

1. Added BACKTRAP define.

2. Added backdoor trapping facility. See notes.

	modified:   linux/network.c
	modified:   linux/sound.c

1. Moved dbg_printf() macro to the include.

	modified:   linux/terminal.c

1. Allow enter key to exit autohold.

	modified:   linux/terminal.not

1. Added terminal notes.

Notes:

The backdoor trapping mode works by giving callers who create files an invalid
address for FILE* that will cause a trap if used. The idea was to use this to
find backdoor accesses into the FILE structure and make stdio work in Linux.

It didn't work, I found that modules were simply bypassing the calls in stdio.
However, the tool was very useful, and I am commiting it so that I have a record
of it. It is not generally applicable, so I am removing it again in the next
commit.