Added prototypes for new functions added in Linux.
 Changes to be committed:
	modified:   windows/terminal.c

1. Added autohold() content, this does nothing on Windows, see notes.

2. Added prototypes for wrtstrn(), sizbuf(), and title(). These are not filled
yet.

Notes:

1. The only tests for the extensions at present are for sizbuf() and the buffer
follow mode, which is allowed to be unimplemented.

2. Using transpartent mode on Windows, it is not required to have autohold in
Windows console. If the screen is not changed, the original screen will be left
as is. If the screen is changed, then clearly the program is terminal aware.

I did a deep dive into why linux does not support transparent mode. It was
declared to be a security risk because any readback of the screen is inserted
into the input buffer, and then can be executed. Windows works differently,
and so does not have that risk. Thus, this appears to be permanent unless a
custom version of the underlying xterm would be used. The pseudo-terminal
device interface would also solve that, but there are other issues with that,
like the need for supervisor access to IOCTL interfaces.