Fix threading issue in thread test.
 Changes to be committed:
	modified:   tests/terminal_test.c

1. Add double buffering to stop flag.

2. Get true state of flag in locked section.

3. In test, move stop flag into lock.

Notes:

A classing threading signal issue. The signal from the subthread will only be
caught if the main thread is already waiting for it when the signal is made.
If the stop flag is set outside of the lock, there is a finite chance that
the subthread will see it, signal and cancel before the main thread enters
the wait. The solution is to move the stop flag into the lock, then the
subthread must lock before sampling the stop flag. This means that if the
main thread has flagged, but not yet entered the wait, the locked status
will hold the subthread until it is.

This is the nature of a single shot semaphore. Note that we have used the
mechanisim to form a set/reset flag, ie., the single shot semaphore can be a
building block for other constructs.