Moved error codes to avoid other module collisions.
 Changes to be committed:
	modified:   include/terminal.h
	modified:   linux/terminal.c

1. Coined the error code names with "disp" (display) to avoid collision with
other modules in the pa system.

notes:

All of the PA modules have an error() routine. To avoid the network module (say)
colliding with the terminal module, we coin them. The display modules, terminal
and graphics, use the same error codes, so the coin is disp (display) to group
them together.