Add terminal notes; Fit mac/system_event for threading; Apply event reset fix from Linux.
 Changes to be committed:
       modified:   linux/system_event.c

1. Move debug print to include.

       modified:   linux/terminal.c

1. Remove obsolete comment.

       modified:   linux/terminal.not

1. Add BSD terminal notes.

       modified:   macosx/system_event.c

1. Add pthreads.

2. Move diagnostic to include file.

3. Add evtlock, thread lock for system event package.

4. Add resetsev, the reset signal system event.

5. Wrap all API routines with locking.

6. Add reset fix from linux.

Notes:

1. This was covered in Linux. After putting events on a thread, the operating
thread calls kevent() and stays there, so if new events are registered, they
are ignored. We fix this by sending a SIGUSR1 signal to the same thread,
which causes kevent() to exit and thus reset the event structures.

This relies on the idea that chgevt is only read one time by kevent().
Note there is a race between the setup for that call and any other routines
modifying the chgevt data structure. I think we need to double buffer the
structure and copy it into place with the lock before calling kevent().

2. The terminal test works on the majority of tests now.

3. system_event should be moved to the BSD area, since Mac OSX was derived from
BSD (and not the other way around).