Use fixed attributes for mutex creation; fix race condition.
 Changes to be committed:
	modified:   Makefile
	modified:   linux/services.c

1. As per recommendation, use specified attributes on mutex creation in
initlock(). This prevents use of recursive locking, which can cause real
problems, and also specifies return of error conditions (the default is to
ignore them? WTF?).

	modified:   tests/terminal_test.c

1. Prevent identified race condition with setting the thread stop flag and
awaiting the thread exit signal. If thdstp is set, the subthread could exit and
signal an exit before the main thread enters waitsig(). This is because the
subthread could be out of lock and exit. Having the subthread lock before
sending the signal does the trick, because waitsig() is by definition within a
lock. This tends to back up the idea that sendsig() should also always be
wrapped in a lock!

Notes:

I have also identified a problem with both threads using waittime(), abet with
different timer ids. The threads consume each other's messages and thus
malfunctions. This will be fixed in a subsequent push.