Use full attribute for pthead_cond_init(); Fix error in waitsig(); Change error title to xterm; Refactor thread test.
The thread test was refactored. A single thread is assigned to read events, and
then it sends signals to the main and subthreads. In this way, only a single
reads from the event queue. I suspect this is the template for how to solve this
kind of issue.

 Changes to be committed:
	modified:   include/terminal.h
	modified:   linux/services.c

1. Use full attributes for pthread_cond_init(). This makes sure we have the
defaults we want.

2. Fix error in waitsig().

	modified:   linux/terminal.c

1. Change error preamble to xterm.

	modified:   tests/terminal_test.c

1. Add eventthread() thread main routine, reads events and dispatches signals
to other threads.

2. Fix thread() thread main routine to actually use the stop flag.

3. Refactor thread() to accept timeout signals.

4. Refactor thread test to accept timeout signals.