Add ability to use clang instead of gcc on linux and windows.
clang can be installed as an alternate compiler in Linux and Windows. It is also
the standard compiler for BSD and Mac OSX. The makefile was updated for this,
and errors/warnings generated by the compiler fixed.

 Changes to be committed:
	modified:   Makefile

1. Introduce overridable variable CLANG.

2. Use CLANG as alternative compiler selector for Windows and Linux.

	modified:   cpp/terminal.cpp
	modified:   hpp/terminal.hpp

1. Added definitions for focus, nofocus, hover and nohover.

	modified:   linux/sound.c

1. Added our (infamous) strcmpus() to allow compare of unsigned strings to
signed constant strings and stop warnings in clang. Used anywhere we are
comparing constants.

	modified:   tests/terminal_test.c
	modified:   tests/terminal_test.not

1. Removed solved threading test issue.