Add lib directory and place all .a and .so files in that.
The "standard" is to use a fork of:

include - Include files.
bin     - Executables.
lib     - Libraries (.a/.so)

To comply, we move the lib files we create to a new library directory.
The only one this does not work for is glibc/libc.so.6, which is tied
into the loader in odd ways.

 Changes to be committed:
	modified:   Makefile

1. Move all .a and .so references in the make to lib directory.